Geo/requirements.txt
numpy
rasterio>=1.3
fiona>=1.9
shapely>=2.0
pyproj>=3.0
requests
psutil
Pillow
